Use u8 type to hold the number of ECC symbols
Previoulsy usize was used. However, I don't think there is a need to use
a target dependent sized type here: the number must always be between 0
and 30 (inclusive) - which fits nicely in the u8 range. Using a u8 may
also allow for some very-minor size optimizations - so it seems like an
overall better fit.

+++ benches/benchmark.rs
@@ -5,7 +5,7 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use reed_solomon_32::{Buffer, correct, encode};
 
 struct TestEncode {
-    ecc: usize,
+    ecc: u8,
     message: [u8; 31],
     message_len: usize,
 }
@@ -15,7 +15,7 @@ fn test_encode(params: TestEncode) -> Buffer {
 }
 
 struct TestDecode {
-    ecc: usize,
+    ecc: u8,
     encoded: [u8; 31],
     encoded_len: usize,
 }
